refactor: 폴더구조 및 파일명 변경, 통신코드 수정
